+++ test/base/list.test.ts
@@ -1,10 +1,14 @@
 import { flatten, flat, shuffle, getAccCellWidth  } from '../../src';
+import * as listUtils from '../../src/base/list';
 
 describe('flatten', () => {
   it('flatten', () => {
     expect(typeof flatten([{ id: 1, name: 'a' }], 'id')).toBe('object');
 
     expect(flatten([{ id: 1, name: 'a' }], 'id')[1].name).toBe('a');
+
+
+    expect(listUtils.flatten([{ id: 1, name: 'a' }], 'id')[1].name).toBe('a');
   });
 });